refactor(blend): use uint32 for Auction ledger fields
StartBlock and LastModifiedLedger were int32 while every other blend row struct
uses uint32 for ledger-valued fields, casting to int32 only at the write boundary.
Align Auction with that convention.

+++ internal/data/blend/auctions.go
@@ -33,8 +33,8 @@ type Auction struct {
 	AuctionType        int32
 	Bid                map[string]string
 	Lot                map[string]string
-	StartBlock         int32
-	LastModifiedLedger int32
+	StartBlock         uint32
+	LastModifiedLedger uint32
 }
 
 // AuctionModelInterface exposes Blend v2 active-auction storage operations.
@@ -106,8 +106,8 @@ func (m *AuctionModel) BatchUpsert(ctx context.Context, dbTx pgx.Tx, rows []Auct
 		auctionTypes[i] = r.AuctionType
 		bids[i] = bidJSON
 		lots[i] = lotJSON
-		startBlocks[i] = r.StartBlock
-		ledgers[i] = r.LastModifiedLedger
+		startBlocks[i] = int32(r.StartBlock)
+		ledgers[i] = int32(r.LastModifiedLedger)
 	}
 
 	const upsertQuery = `
